progress on an initiative view
